mv flattens from pytrees to optimize, marked private
