+++ pennylane/optimize/rotoselect.py
@@ -16,7 +16,7 @@
 import numpy as np
 
 import pennylane as qml
-from pennylane.pytrees import flatten_np, unflatten_np
+from .qng import _flatten_np, _unflatten_np
 
 
 class RotoselectOptimizer:
@@ -132,11 +132,11 @@ class RotoselectOptimizer:
         Returns:
             array: The new variable values :math:`x^{(t+1)}` as well as the new generators.
         """
-        x_flat = np.fromiter(flatten_np(x), dtype=float)
+        x_flat = np.fromiter(_flatten_np(x), dtype=float)
         # wrap the objective function so that it accepts the flattened parameter array
         # pylint:disable=unnecessary-lambda-assignment
         objective_fn_flat = lambda x_flat, gen: objective_fn(
-            unflatten_np(x_flat, x), generators=gen, **kwargs
+            _unflatten_np(x_flat, x), generators=gen, **kwargs
         )
 
         try:
@@ -151,7 +151,7 @@ class RotoselectOptimizer:
                 objective_fn_flat, x_flat, generators, d
             )
 
-        return unflatten_np(x_flat, x), generators
+        return _unflatten_np(x_flat, x), generators
 
     def _find_optimal_generators(self, objective_fn, x, generators, d):
         r"""Optimizer for the generators.
