+++ pennylane/pytrees/pytrees.py
@@ -296,78 +296,3 @@ def _unflatten(new_data, structure):
         return next(new_data)
     children = tuple(_unflatten(new_data, s) for s in structure.children)
     return unflatten_registrations[structure.type_](children, structure.metadata)
-
-
-def flatten_np(x):
-    """Iterate recursively through an arbitrarily nested structure in depth-first order.
-
-    See also :func:`_unflatten`.
-
-    Args:
-        x (array, Iterable, Any): each element of an array or an Iterable may itself be any of these types
-
-    Yields:
-        Any: elements of x in depth-first order
-    """
-    if isinstance(x, np.ndarray):
-        yield from flatten_np(x.flat)  # should we allow object arrays? or just "yield from x.flat"?
-    elif isinstance(x, qml.wires.Wires):
-        # Reursive calls to flatten `Wires` will cause infinite recursion (`Wires` atoms are `Wires`).
-        # Since Wires are always flat, just yield.
-        yield from x
-    elif isinstance(x, Iterable) and not isinstance(x, (str, bytes)):
-        for item in x:
-            yield from flatten_np(item)
-    else:
-        yield x
-
-
-def _unflatten_np(flat, model):
-    """Restores an arbitrary nested structure to a flattened iterable.
-
-    See also :func:`_flatten`.
-
-    Args:
-        flat (array): 1D array of items
-        model (array, Iterable, Number): model nested structure
-
-    Raises:
-        TypeError: if ``model`` contains an object of unsupported type
-
-    Returns:
-        Union[array, list, Any], array: first elements of flat arranged into the nested
-        structure of model, unused elements of flat
-    """
-    if isinstance(model, (numbers.Number, str)):
-        return flat[0], flat[1:]
-
-    if isinstance(model, np.ndarray):
-        idx = model.size
-        res = np.array(flat)[:idx].reshape(model.shape)
-        return res, flat[idx:]
-
-    if isinstance(model, Iterable):
-        res = []
-        for x in model:
-            val, flat = _unflatten_np(flat, x)
-            res.append(val)
-        return res, flat
-
-    raise TypeError(f"Unsupported type in the model: {type(model)}")
-
-
-def unflatten_np(flat, model):
-    """Wrapper for :func:`_unflatten`.
-
-    Args:
-        flat (array): 1D array of items
-        model (array, Iterable, Number): model nested structure
-
-    Raises:
-        ValueError: if ``flat`` has more elements than ``model``
-    """
-    # pylint:disable=len-as-condition
-    res, tail = _unflatten_np(np.asarray(flat), model)
-    if len(tail) != 0:
-        raise ValueError("Flattened iterable has more elements than the model.")
-    return res
